It is a 64bit machine. Thus hash should return long int.

+++ sym.c
@@ -172,9 +172,9 @@ char *slookup(char str[])
 /*
  * hash - hash name, turn address into hash number
  */
-int hash(char *s)
+long hash(char *s)
 {
-   return((int ) s);
+   return((long ) s);
 }   
 
 /*
